modified:   draft.ipynb 	modified:   experiments.ipynb 	modified:   full_finetuning.py 	deleted:    inst_1000_xstest.csv 	deleted:    inst_1000_xstest_list.csv 	deleted:    inst_2000_xstest.csv 	deleted:    inst_2000_xstest_list.csv 	deleted:    inst_500_xstest.csv 	deleted:    inst_500_xstest_list.csv 	modified:   run.sh

+++ full_finetuning.py
@@ -221,7 +221,7 @@ def train(
     model = AutoModelForCausalLM.from_pretrained(
         base_model,
         load_in_8bit=True if train_using_lora else False,
-        torch_dtype=torch.float16 if train_using_lora else torch.bfloat16,
+        torch_dtype=torch.float16 if train_using_lora else torch.float16,
         device_map=device_map,
         trust_remote_code=True,
     )
